+++ JS/ES-VE/economic_data.js
@@ -273,4 +273,7 @@ $('#streetsdatatable').DataTable().draw();
 
 //reloads the property selectors
 updatepropsels();
+
+//shows the current player's transaction data
+loadplayeropts();
 //-------------------------------------------------------------